5.7,运动数据叠加，修复bug 4.19,基本功能初步完成 4.11，历史记录、我的报告完善 4.8，跑步模块修改，分析中加入运动，代码整合 3.24 3.22,完善报告、分析数据 3.19,数据上传 2.28，演示结束后 1.19,年前工作结束 1.16,社区基本完善 1.14,社区界面 1.9,社区开始 2017.1.5,首页修改 12.30，元旦之前 12.28，分析报告 12.23,周五 12.22 12.21 12.20,健康计划 12.16 12.15 12.13 12.12中午，晚上 12.9电脑坏之前 12.8心电图 11.29,注册 11.28,分析图做了大概 11.25, 11-24, 2016-11-23，初始化

+++ app/src/main/java/com/amsu/healthy/bean/FullReport.java
@@ -17,7 +17,7 @@ public class FullReport {
         List<ReportItem> HRRrep;
         List<ReportItem> HRVrep;*/
         public List<HRrepBean> HRrep;
-        public int[] ECrep;
+        public float[] ECrep;
         public List<HRRrepBean> HRRrep;
         public List<HRVrepBean> HRVrep;
         public List<List<String>> HRlist;
